move testcontainer set up to an extension. update the parallelism config.
