+++ core/src/test/java/dev/morphia/test/TemplatedTestBase.java
@@ -57,8 +57,6 @@ public abstract class TemplatedTestBase extends TestBase {
     protected final ObjectMapper mapper = new ObjectMapper();
 
     public TemplatedTestBase() {
-        buildConfig()
-                .codecProvider(new ZDTCodecProvider());
     }
 
     protected static String toString(List<Document> actual, String prefix) {
