+++ pom.xml
@@ -131,7 +131,7 @@
                     <artifactId>maven-surefire-plugin</artifactId>
                     <configuration>
                         <useModulePath>false</useModulePath>
-                        <forkCount>1C</forkCount>
+                        <forkCount>1</forkCount>
                         <systemPropertyVariables>
                             <testcontainers.reuse.enable>true</testcontainers.reuse.enable>
                             <junit.jupiter.testinstance.lifecycle.default>per_class</junit.jupiter.testinstance.lifecycle.default>
